renamed files to comply with qubic style
