+++ asset_utils.cpp
@@ -2,15 +2,15 @@
 #include <cstring>
 
 #include "structs.h"
-#include "walletUtils.h"
-#include "keyUtils.h"
-#include "assetUtil.h"
+#include "wallet_utils.h"
+#include "key_utils.h"
+#include "asset_utils.h"
 #include "connection.h"
 #include "logger.h"
-#include "nodeUtils.h"
-#include "K12AndKeyUtil.h"
+#include "node_utils.h"
+#include "k12_and_key_utils.h"
 #include "utils.h"
-#include "sanityCheck.h"
+#include "sanity_check.h"
 
 std::vector<RespondOwnedAssets> getOwnedAsset(const char * nodeIp, const int nodePort, const char* requestedIdentity)
 {
