+++ asset_utils.h
