+++ common_functions.h
