--- fourq-qubic.cpp
@@ -1,2 +1,0 @@
-#include "fourq-qubic.h"
-#include "K12AndKeyUtil.h"
