+++ fourq_qubic.h
