+++ node_utils.h
