+++ qvault.cpp
@@ -3,13 +3,13 @@
 #include <stdexcept>
 
 #include "structs.h"
-#include "walletUtils.h"
-#include "keyUtils.h"
-#include "assetUtil.h"
+#include "wallet_utils.h"
+#include "key_utils.h"
+#include "asset_utils.h"
 #include "connection.h"
 #include "logger.h"
-#include "nodeUtils.h"
-#include "K12AndKeyUtil.h"
+#include "node_utils.h"
+#include "k12_and_key_utils.h"
 #include "qvault.h"
 
 #define QVAULT_CONTRACT_INDEX 10
