+++ wallet_utils.h
